# [4.0] String performance optimizations