--- a/stdlib/public/core/UTFEncoding.swift
+++ b/stdlib/public/core/UTFEncoding.swift
@@ -20,13 +20,13 @@ public protocol _UTFParser {
   associatedtype Encoding : _UnicodeEncoding_
 
   func _parseMultipleCodeUnits() -> (isValid: Bool, bitCount: UInt8)
-  func _bufferedScalar(bitCount: UInt8) -> _UIntBuffer<UInt32, Encoding.CodeUnit>
+  func _bufferedScalar(bitCount: UInt8) -> Encoding.EncodedScalar
   
   var _buffer: _UIntBuffer<UInt32, Encoding.CodeUnit> { get set }
 }
 
 extension _UTFParser
-where Encoding.EncodedScalar == _UIntBuffer<UInt32, Encoding.CodeUnit> {
+where Encoding.EncodedScalar : RangeReplaceableCollection {
 
   @inline(__always)
   public mutating func parseScalar<I : IteratorProtocol>(
@@ -39,7 +39,7 @@ where Encoding.EncodedScalar == _UIntBuffer<UInt32, Encoding.CodeUnit> {
       guard let codeUnit = input.next() else { return .emptyInput }
       // ASCII, return immediately.
       if Encoding._isScalar(codeUnit) {
-        return .valid(Encoding.EncodedScalar(containing: codeUnit))
+        return .valid(Encoding.EncodedScalar(CollectionOfOne(codeUnit)))
       }
       // Non-ASCII, proceed to buffering mode.
       _buffer.append(codeUnit)
@@ -50,7 +50,7 @@ where Encoding.EncodedScalar == _UIntBuffer<UInt32, Encoding.CodeUnit> {
       // to bufferless mode once we've exhausted it.
       let codeUnit = Encoding.CodeUnit(extendingOrTruncating: _buffer._storage)
       _buffer.remove(at: _buffer.startIndex)
-      return .valid(Encoding.EncodedScalar(containing: codeUnit))
+      return .valid(Encoding.EncodedScalar(CollectionOfOne(codeUnit)))
     }
     // Buffering mode.
     // Fill buffer back to 4 bytes (or as many as are left in the iterator).