Explanation: String optimizations
Scope: optimizations
Radar: rdar://32744663 String Optimizations for Swift 4 beta 3
Risk: very low
Testing: Passed full validation + benchmarking on master; validation and source compatibility on the 4.0 branch (benchmark testing on 4.0 branch not supported)