@dabrahams 2017-06-13
@swift-ci Please test source compatibility