@shahmishal 2017-06-13
Benchmark driver is written to only work with master branch 

https://github.com/apple/swift/blob/swift-4.0-branch/benchmark/scripts/Benchmark_Driver#L292

```
if current_branch != 'master' and not os.path.isdir(master_branch_dir):
        print('Unable to find benchmark logs for master branch. Set a ' +
              'baseline benchmark log by passing --benchmark to ' +
              'build-script while on master branch.')
        return 1
```