@dabrahams 2017-06-14
Locally-derived benchmark results:

<details open>
  <summary>Regression (10)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP 
---                                                     | ---    | ---    | ---    | ---     
ObserverForwarderStruct                                 | 971    | 1191   | +22.7% | **0.82x** 
ProtocolDispatch                                        | 2153   | 2626   | +22.0% | **0.82x** 
StringBuilderLong                                       | 804    | 941    | +17.0% | **0.85x** 
MonteCarloE                                             | 8874   | 9866   | +11.2% | **0.90x** 
ObjectiveCBridgeStubDateAccess                          | 188    | 209    | +11.2% | **0.90x** 
Memset                                                  | 178    | 195    | +9.6%  | **0.91x** 
Array2D                                                 | 1684   | 1839   | +9.2%  | **0.92x** 
Calculator                                              | 29     | 31     | +6.9%  | **0.94x** 
MapReduceString                                         | 90     | 96     | +6.7%  | **0.94x** 
DropWhileAnyCollectionLazy                              | 92     | 98     | +6.5%  | **0.94x** 

</details>

<details >
  <summary>Improvement (57)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP 
---                                                     | ---    | ---    | ---    | ---     
Chars                                                   | 1835   | 732    | -60.1% | **2.51x** 
SuffixArrayLazy                                         | 10     | 6      | -40.0% | **1.67x** 
SuffixArray                                             | 10     | 6      | -40.0% | **1.67x** 
SuperChars                                              | 163185 | 130367 | -20.1% | **1.25x** 
DropWhileCountableRange                                 | 25     | 20     | -20.0% | **1.25x** 
StringWalk                                              | 1108   | 889    | -19.8% | **1.25x** 
StrComplexWalk                                          | 737    | 608    | -17.5% | **1.21x** 
LazilyFilteredArrays                                    | 56308  | 46857  | -16.8% | **1.20x** 
PrefixSequenceLazy                                      | 1637   | 1386   | -15.3% | **1.18x** 
ObjectiveCBridgeFromNSStringForced                      | 2075   | 1817   | -12.4% | **1.14x** 
OpenClose                                               | 50     | 44     | -12.0% | **1.14x** 
SuffixSequenceLazy                                      | 4411   | 3883   | -12.0% | **1.14x** 
ObjectiveCBridgeStubFromArrayOfNSString                 | 25506  | 22570  | -11.5% | **1.13x** 
SuffixCountableRangeLazy                                | 9      | 8      | -11.1% | **1.12x** 
SuffixCountableRange                                    | 9      | 8      | -11.1% | **1.12x** 
DropLastCountableRangeLazy                              | 9      | 8      | -11.1% | **1.12x** 
ObjectiveCBridgeFromNSSetAnyObjectToStringForced        | 57533  | 51447  | -10.6% | **1.12x** 
PrefixWhileSequenceLazy                                 | 29     | 26     | -10.3% | **1.12x** 
StringWithCString                                       | 55048  | 49505  | -10.1% | **1.11x** 
PolymorphicCalls                                        | 20     | 18     | -10.0% | **1.11x** 
PrefixSequence                                          | 1569   | 1417   | -9.7%  | **1.11x** 
NSDictionaryCastToSwift                                 | 4937   | 4489   | -9.1%  | **1.10x** 
ObjectAllocation                                        | 171    | 157    | -8.2%  | **1.09x** 
CharacterLiteralsSmall                                  | 315    | 290    | -7.9%  | **1.09x** 
ObjectiveCBridgeToNSArray                               | 25782  | 23777  | -7.8%  | **1.08x** 
BitCount                                                | 13     | 12     | -7.7%  | **1.08x** 
DropLastSequenceLazy                                    | 446    | 412    | -7.6%  | **1.08x** 
Walsh                                                   | 299    | 277    | -7.4%  | **1.08x** 
SubstringEqualString                                    | 3103   | 2878   | -7.3%  | **1.08x** 
SetExclusiveOr                                          | 2689   | 2495   | -7.2%  | **1.08x** 
PrefixWhileAnySeqCRangeIterLazy                         | 70     | 65     | -7.1%  | **1.08x** 
PrefixWhileCountableRangeLazy                           | 28     | 26     | -7.1%  | **1.08x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectForced         | 4707   | 4383   | -6.9%  | **1.07x** 
ArrayOfPOD                                              | 146    | 136    | -6.8%  | **1.07x** 
SuffixAnySeqCntRange                                    | 15     | 14     | -6.7%  | **1.07x** 
DropLastSequence                                        | 441    | 412    | -6.6%  | **1.07x** 
ObjectiveCBridgeFromNSArrayAnyObjectForced              | 3937   | 3686   | -6.4%  | **1.07x** 
PrefixWhileAnySeqCRangeIter                             | 11715  | 10977  | -6.3%  | **1.07x** 
ReversedArray                                           | 48     | 45     | -6.2%  | **1.07x** 
StringHasPrefix                                         | 16     | 15     | -6.2%  | **1.07x** 
NopDeinit                                               | 24934  | 23442  | -6.0%  | **1.06x** 
ObjectiveCBridgeFromNSArrayAnyObjectToStringForced      | 32537  | 30596  | -6.0%  | **1.06x** 
MapReduce                                               | 253    | 238    | -5.9%  | **1.06x** 
StaticArray                                             | 17     | 16     | -5.9%  | **1.06x** 
SortStringsUnicode                                      | 19333  | 18238  | -5.7%  | **1.06x** 
SubstringEquatable                                      | 6608   | 6235   | -5.6%  | **1.06x** 
PrefixWhileArray                                        | 55     | 52     | -5.5%  | **1.06x** 
SortSortedStrings                                       | 779    | 739    | -5.1%  | **1.05x** 
ObjectiveCBridgeStubDataAppend                          | 3588   | 3404   | -5.1%  | **1.05x** 
DropWhileSequenceLazy                                   | 59     | 56     | -5.1%  | **1.05x** 
SevenBoom                                               | 1325   | 1258   | -5.1%  | **1.05x** 
SuffixAnyCollection                                     | 20     | 19     | -5.0%  | **1.05x** 
DropFirstArray                                          | 20     | 19     | -5.0%  | **1.05x** 
DropLastAnyCollection                                   | 20     | 19     | -5.0%  | **1.05x** 
DropFirstArrayLazy                                      | 20     | 19     | -5.0%  | **1.05x** 
StringFromLongWholeSubstringGeneric                     | 20     | 19     | -5.0%  | **1.05x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectToStringForced | 81209  | 77173  | -5.0%  | **1.05x** 

</details>

<details >
  <summary>No Changes (214)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP 
---                                                     | ---    | ---    | ---    | ---     
AngryPhonebook                                          | 2593   | 2706   | +4.4%  | 0.96x   
AnyHashableWithAClass                                   | 58659  | 59777  | +1.9%  | 0.98x   
ArrayAppend                                             | 495    | 489    | -1.2%  | 1.01x   
ArrayAppendArrayOfInt                                   | 357    | 366    | +2.5%  | 0.98x   
ArrayAppendAscii                                        | 17795  | 17789  | -0.0%  | 1.00x   
ArrayAppendFromGeneric                                  | 348    | 352    | +1.1%  | 0.99x   
ArrayAppendGenericStructs                               | 723    | 701    | -3.0%  | 1.03x   
ArrayAppendLatin1                                       | 32570  | 32020  | -1.7%  | 1.02x   
ArrayAppendLazyMap                                      | 716    | 715    | -0.1%  | 1.00x   
ArrayAppendOptionals                                    | 702    | 695    | -1.0%  | 1.01x   
ArrayAppendRepeatCol                                    | 622    | 621    | -0.2%  | 1.00x   
ArrayAppendReserved                                     | 334    | 334    | +0.0%  | 1.00x   
ArrayAppendSequence                                     | 675    | 682    | +1.0%  | 0.99x   
ArrayAppendStrings                                      | 13674  | 13583  | -0.7%  | 1.01x   
ArrayAppendToFromGeneric                                | 351    | 355    | +1.1%  | 0.99x   
ArrayAppendToGeneric                                    | 348    | 356    | +2.3%  | 0.98x   
ArrayAppendUTF16                                        | 30538  | 30588  | +0.2%  | 1.00x   
ArrayInClass                                            | 54     | 55     | +1.9%  | 0.98x   
ArrayLiteral                                            | 1069   | 1075   | +0.6%  | 0.99x   
ArrayOfGenericPOD                                       | 191    | 188    | -1.6%  | 1.02x   
ArrayOfGenericRef                                       | 3539   | 3603   | +1.8%  | 0.98x   
ArrayOfRef                                              | 3731   | 3592   | -3.7%  | 1.04x   
ArrayPlusEqualArrayOfInt                                | 352    | 355    | +0.9%  | 0.99x   
ArrayPlusEqualFiveElementCollection                     | 3731   | 3617   | -3.1%  | 1.03x   
ArrayPlusEqualSingleElementCollection                   | 482    | 493    | +2.3%  | 0.98x   
ArrayPlusEqualThreeElements                             | 1395   | 1386   | -0.6%  | 1.01x   
ArraySubscript                                          | 1309   | 1334   | +1.9%  | 0.98x   
ArrayValueProp                                          | 5      | 5      | +0.0%  | 1.00x   
ArrayValueProp2                                         | 5      | 5      | +0.0%  | 1.00x   
ArrayValueProp3                                         | 5      | 5      | +0.0%  | 1.00x   
ArrayValueProp4                                         | 5      | 5      | +0.0%  | 1.00x   
ByteSwap                                                | 0      | 0      | +0.0%  | 1.00x   
CStringLongAscii                                        | 4003   | 4101   | +2.4%  | 0.98x   
CStringLongNonAscii                                     | 1689   | 1750   | +3.6%  | 0.97x   
CStringShortAscii                                       | 4867   | 4760   | -2.2%  | 1.02x   
CaptureProp                                             | 4020   | 4115   | +2.4%  | 0.98x   
CharacterLiteralsLarge                                  | 4950   | 5098   | +3.0%  | 0.97x   
ClassArrayGetter                                        | 11     | 11     | +0.0%  | 1.00x   
DeadArray                                               | 132    | 130    | -1.5%  | 1.02x   
Dictionary                                              | 469    | 451    | -3.8%  | 1.04x   
Dictionary2                                             | 1549   | 1534   | -1.0%  | 1.01x   
Dictionary2OfObjects                                    | 2833   | 2803   | -1.1%  | 1.01x   
Dictionary3                                             | 411    | 405    | -1.5%  | 1.01x   
Dictionary3OfObjects                                    | 787    | 774    | -1.7%  | 1.02x   
DictionaryBridge                                        | 2702   | 2587   | -4.3%  | 1.04x   
DictionaryLiteral                                       | 1265   | 1253   | -0.9%  | 1.01x   
DictionaryOfObjects                                     | 2052   | 1977   | -3.7%  | 1.04x   
DictionaryRemove                                        | 2013   | 1934   | -3.9%  | 1.04x   
DictionaryRemoveOfObjects                               | 20156  | 20100  | -0.3%  | 1.00x   
DictionarySwap                                          | 329    | 328    | -0.3%  | 1.00x   
DictionarySwapOfObjects                                 | 6037   | 5992   | -0.7%  | 1.01x   
DropFirstAnyCollection                                  | 51     | 50     | -2.0%  | 1.02x   
DropFirstAnyCollectionLazy                              | 41276  | 40387  | -2.2%  | 1.02x   
DropFirstAnySeqCRangeIter                               | 23838  | 23809  | -0.1%  | 1.00x   
DropFirstAnySeqCRangeIterLazy                           | 23776  | 23981  | +0.9%  | 0.99x   
DropFirstAnySeqCntRange                                 | 45     | 43     | -4.4%  | 1.05x   
DropFirstAnySeqCntRangeLazy                             | 44     | 43     | -2.3%  | 1.02x   
DropFirstAnySequence                                    | 5120   | 5068   | -1.0%  | 1.01x   
DropFirstAnySequenceLazy                                | 5093   | 5054   | -0.8%  | 1.01x   
DropFirstCountableRange                                 | 27     | 26     | -3.7%  | 1.04x   
DropFirstCountableRangeLazy                             | 26     | 26     | +0.0%  | 1.00x   
DropFirstSequence                                       | 1914   | 1887   | -1.4%  | 1.01x   
DropFirstSequenceLazy                                   | 1907   | 1831   | -4.0%  | 1.04x   
DropLastAnyCollectionLazy                               | 14007  | 13541  | -3.3%  | 1.03x   
DropLastAnySeqCRangeIter                                | 4143   | 4079   | -1.5%  | 1.02x   
DropLastAnySeqCRangeIterLazy                            | 4153   | 4117   | -0.9%  | 1.01x   
DropLastAnySeqCntRange                                  | 14     | 14     | +0.0%  | 1.00x   
DropLastAnySeqCntRangeLazy                              | 14     | 14     | +0.0%  | 1.00x   
DropLastAnySequence                                     | 5976   | 5845   | -2.2%  | 1.02x   
DropLastAnySequenceLazy                                 | 5924   | 5816   | -1.8%  | 1.02x   
DropLastArray                                           | 6      | 6      | +0.0%  | 1.00x   
DropLastArrayLazy                                       | 6      | 6      | +0.0%  | 1.00x   
DropLastCountableRange                                  | 8      | 8      | +0.0%  | 1.00x   
DropWhileAnyCollection                                  | 63     | 60     | -4.8%  | 1.05x   
DropWhileAnySeqCRangeIter                               | 19284  | 18567  | -3.7%  | 1.04x   
DropWhileAnySeqCRangeIterLazy                           | 91     | 90     | -1.1%  | 1.01x   
DropWhileAnySeqCntRange                                 | 57     | 55     | -3.5%  | 1.04x   
DropWhileAnySeqCntRangeLazy                             | 92     | 90     | -2.2%  | 1.02x   
DropWhileAnySequence                                    | 5832   | 5722   | -1.9%  | 1.02x   
DropWhileAnySequenceLazy                                | 1872   | 1838   | -1.8%  | 1.02x   
DropWhileArray                                          | 31     | 32     | +3.2%  | 0.97x   
DropWhileArrayLazy                                      | 68     | 70     | +2.9%  | 0.97x   
DropWhileCountableRangeLazy                             | 60     | 59     | -1.7%  | 1.02x   
DropWhileSequence                                       | 1426   | 1400   | -1.8%  | 1.02x   
EqualStringSubstring                                    | 421    | 427    | +1.4%  | 0.99x   
EqualSubstringString                                    | 411    | 405    | -1.5%  | 1.01x   
EqualSubstringSubstring                                 | 416    | 404    | -2.9%  | 1.03x   
EqualSubstringSubstringGenericEquatable                 | 414    | 404    | -2.4%  | 1.02x   
ErrorHandling                                           | 2460   | 2393   | -2.7%  | 1.03x   
GlobalClass                                             | 0      | 0      | +0.0%  | 1.00x   
Hanoi                                                   | 3017   | 3049   | +1.1%  | 0.99x   
HashTest                                                | 1394   | 1438   | +3.2%  | 0.97x   
Histogram                                               | 233    | 227    | -2.6%  | 1.03x   
Integrate                                               | 208    | 202    | -2.9%  | 1.03x   
IterateData                                             | 638    | 616    | -3.4%  | 1.04x   
Join                                                    | 340    | 325    | -4.4%  | 1.05x   
LazilyFilteredRange                                     | 2867   | 2799   | -2.4%  | 1.02x   
LessSubstringSubstring                                  | 408    | 408    | +0.0%  | 1.00x   
LessSubstringSubstringGenericComparable                 | 420    | 400    | -4.8%  | 1.05x   
LinkedList                                              | 6329   | 6238   | -1.4%  | 1.01x   
MapReduceAnyCollection                                  | 254    | 242    | -4.7%  | 1.05x   
MapReduceAnyCollectionShort                             | 1777   | 1731   | -2.6%  | 1.03x   
MapReduceClass                                          | 2709   | 2646   | -2.3%  | 1.02x   
MapReduceClassShort                                     | 4041   | 3902   | -3.4%  | 1.04x   
MapReduceLazyCollection                                 | 14     | 14     | +0.0%  | 1.00x   
MapReduceLazyCollectionShort                            | 37     | 38     | +2.7%  | 0.97x   
MapReduceLazySequence                                   | 79     | 81     | +2.5%  | 0.98x   
MapReduceSequence                                       | 401    | 406    | +1.2%  | 0.99x   
MapReduceShort                                          | 1729   | 1692   | -2.1%  | 1.02x   
MapReduceShortString                                    | 19     | 19     | +0.0%  | 1.00x   
MonteCarloPi                                            | 38569  | 37610  | -2.5%  | 1.03x   
NSError                                                 | 267    | 256    | -4.1%  | 1.04x   
NSStringConversion                                      | 602    | 595    | -1.2%  | 1.01x   
ObjectiveCBridgeFromNSArrayAnyObject                    | 20220  | 20098  | -0.6%  | 1.01x   
ObjectiveCBridgeFromNSArrayAnyObjectToString            | 37211  | 36435  | -2.1%  | 1.02x   
ObjectiveCBridgeFromNSDictionaryAnyObject               | 102451 | 101435 | -1.0%  | 1.01x   
ObjectiveCBridgeFromNSDictionaryAnyObjectToString       | 88508  | 85787  | -3.1%  | 1.03x   
ObjectiveCBridgeFromNSSetAnyObject                      | 54865  | 56480  | +2.9%  | 0.97x   
ObjectiveCBridgeFromNSSetAnyObjectForced                | 3398   | 3476   | +2.3%  | 0.98x   
ObjectiveCBridgeFromNSSetAnyObjectToString              | 62533  | 59773  | -4.4%  | 1.05x   
ObjectiveCBridgeFromNSString                            | 1305   | 1272   | -2.5%  | 1.03x   
ObjectiveCBridgeStubDateMutation                        | 243    | 235    | -3.3%  | 1.03x   
ObjectiveCBridgeStubFromNSDate                          | 2945   | 2829   | -3.9%  | 1.04x   
ObjectiveCBridgeStubFromNSDateRef                       | 3078   | 2988   | -2.9%  | 1.03x   
ObjectiveCBridgeStubFromNSString                        | 743    | 711    | -4.3%  | 1.05x   
ObjectiveCBridgeStubFromNSStringRef                     | 137    | 131    | -4.4%  | 1.05x   
ObjectiveCBridgeStubNSDataAppend                        | 2143   | 2099   | -2.1%  | 1.02x   
ObjectiveCBridgeStubNSDateMutationRef                   | 10667  | 10951  | +2.7%  | 0.97x   
ObjectiveCBridgeStubNSDateRefAccess                     | 278    | 268    | -3.6%  | 1.04x   
ObjectiveCBridgeStubToArrayOfNSString                   | 25020  | 24772  | -1.0%  | 1.01x   
ObjectiveCBridgeStubToNSDate                            | 13323  | 13239  | -0.6%  | 1.01x   
ObjectiveCBridgeStubToNSDateRef                         | 2894   | 2806   | -3.0%  | 1.03x   
ObjectiveCBridgeStubToNSString                          | 1304   | 1272   | -2.5%  | 1.03x   
…